interface/requirements.txt
pyserial
PyQt6
pyqt6-tools